pyproject.toml
[tool.poetry]
name = "repl_python3_messibot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
flask = "^1.1.2"
replit = "^3.0.0"
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
